Clang tidy of entire codebase

+++ src/Systems/GameplaySys.cpp
@@ -155,8 +155,8 @@ void GameplaySys::update(float deltaTime) {
 
     // Random debug stuff
     PropEntity::instances[1].get<PositionComp>().coords.x =
-        10 + 2 * sin(SDL_GetTicks() / 1000.0f);
-    PropEntity::instances[2].get<PositionComp>().rotation.x = SDL_GetTicks() / 1000.0f;
+        10 + 2 * sin(static_cast<float>(SDL_GetTicks()) / 1000.0f);
+    PropEntity::instances[2].get<PositionComp>().rotation.x = static_cast<float>(SDL_GetTicks()) / 1000.0f;
 
     // PropEntity::instances[0].get<PositionComp>().coords.x = playerPosition.x;
     // PropEntity::instances[0].get<PositionComp>().coords.y = playerPosition.y;
